+++ doc/release-notes.md
@@ -40,7 +40,9 @@ this change in requirements will have minimal to zero impact on our users, most 
 
 ## New RPC methods
 
-TODO
+- The `getindexinfo` RPC returns the actively running indices of the node,
+  including their current sync status and height. It also accepts an `index_name`
+  to specify returning only the status of that index.
 
 ## User interface changes